% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin3}
\alias{admin3}
\title{Vietnamese admin3 dictionary}
\format{An object of class \code{character} of length 80621.}
\usage{
data(admin3)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the admin3 names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
