% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communes}
\alias{communes}
\title{Vietnamese Communes dictionary}
\format{An object of class \code{character} of length 80621.}
\usage{
data(communes)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the communes names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
