% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Retrieve the dictionary definition of an English word}
\usage{
define(word)
}
\arguments{
\item{word}{The word you wish to define}
}
\value{
A tibble with a row for each unique combination of word,
    meaning_number, part_of_speech, phonetic, audio, origin, definition,
    example, synonyms, and antonyms.
}
\description{
Retrieve the dictionary definition of an English word
}
\examples{
\dontrun{
  # Get the definition of "hello"
  define("hello") # Returns a 3 row data.frame

  # Homonyms (words with >1 meaning)
  define("bark") # Returns a 9 row data.frame

  # Unknown words will return a 0 row data.frame and message
  define("sdfsdfsdfsdfsdf")
  # No definition found for sdfsdfsdfsdfsdf

}

}
