% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{package}
\name{dictionary}
\alias{dictionary}
\alias{dictionary-package}
\title{dictionary: Dictionary of Vietnamese admin1 Names}
\description{
This package contains a dictionary for Vietnamese admin1 names, admin2
names and admin3 names, useful to ensure consistency between different data
 packages using vietnamese admin1, districts and/or admin3.
}
