% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{Districts dictionary}
\format{An object of class \code{character} of length 3406.}
\usage{
data(districts)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the districts names in Vietnamese.
}
\keyword{datasets}
