% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO_admin1}
\alias{ISO_admin1}
\title{ISO admin1 unit dictionary}
\format{An object of class \code{character} of length 2357.}
\usage{
data(ISO_admin1)
}
\description{
A named character vector containing the translation in ISO code 3166-66 of the
admin1 unit names of South East Asia country.
}
\keyword{datasets}
