% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO_country}
\alias{ISO_country}
\title{ISO Countries dictionary}
\format{An object of class \code{character} of length 41.}
\usage{
data(ISO_country)
}
\description{
A named character vector containing the translation in ISO code 3166-66 of the
country names of South East Asia.
}
\keyword{datasets}
