% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO_province}
\alias{ISO_province}
\title{ISO Provinces dictionary}
\format{An object of class \code{character} of length 2357.}
\usage{
data(ISO_province)
}
\description{
A named character vector containing the translation in ISO code 3166-66 of the
provinces names of South East Asia country.
}
\keyword{datasets}
