% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kh_admin1}
\alias{kh_admin1}
\title{Cambodian admin1 unit dictionary}
\format{An object of class \code{character} of length 3812.}
\usage{
data(kh_admin1)
}
\description{
A named character vector containing the translation in English of the
admin1 unit names in Cambodia.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
