% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kh_admin1_year}
\alias{kh_admin1_year}
\title{Cambodian admin1 unit By Year}
\format{An object of class \code{list} of length 3.
 \itemize{
   \item \code{"1994-1997"}: character vector of the 23 admin1 unit in English
   \item \code{"1997-2013"}: character vector of the 24 admin1 unit in English
   \item \code{"2013-2020"}: character vector of the 25 admin1 unit in English
}}
\usage{
data(kh_admin1_year)
}
\description{
A list of charactor vector containing the English names of the admin1 unit
ordered by year. This list contains for each of change, the name of the "new"
admin1 unit in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(kh_admin1_year)

# To have the names of the admin1, between 1992 and 1997:
kh_admin1_year$`1994`

}
\keyword{datasets}
