% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kh_province}
\alias{kh_province}
\title{Cambodian Provinces dictionary}
\format{An object of class \code{character} of length 3812.}
\usage{
data(kh_province)
}
\description{
A named character vector containing the translation in English of the
provinces names in Cambodia.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
