% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_admin1}
\alias{la_admin1}
\title{Lao admin1 unit dictionary}
\format{An object of class \code{character} of length 1353.}
\usage{
data(la_admin1)
}
\description{
A named character vector containing the translation in English of the
admin1 unit names in Laos.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
