% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_admin1_year}
\alias{la_admin1_year}
\title{Lao admin1 unit By Year}
\format{An object of class \code{list} of length 3.
 \itemize{
   \item \code{"1997-2006"}: character vector of the 18 admin1 unit in English
   \item \code{"2006-2013"}: character vector of the 17 admin1 unit in English
   \item \code{"2013-2020"}: character vector of the 18 admin1 unit in English
}}
\usage{
data(la_admin1_year)
}
\description{
A list of charactor vector containing the English names of the admin1 unit
ordered by year. This list contains for each of change, the name of the "new"
admin1 unit in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(la_admin1_year)

# To have the names of the admin1, between 1997 and 2006:
la_admin1_year$`1997`

}
\keyword{datasets}
