% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_admin2}
\alias{la_admin2}
\title{Lao admin2 units dictionary}
\format{An object of class \code{character} of length 1632.}
\usage{
data(la_admin2)
}
\description{
A names charactor vector containing the translation from Lao to
English of the admin2 units names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
