% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_district}
\alias{la_district}
\title{Lao Districts dictionary}
\format{An object of class \code{character} of length 1626.}
\usage{
data(la_district)
}
\description{
A names charactor vector containing the translation from Lao to
English of the districts names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
