% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_history}
\alias{la_history}
\title{Lao admin1 unit Boundaries History (since 1993)}
\format{An object of class \code{list} of length 2.}
\usage{
data(la_history)
}
\description{
List of 2 lists containing 6 elements:
 \itemize{
   \item \code{year}: date of event in character in format "YYYY-mm-dd"
   \item \code{event}: character vector of one object either split, merge,
   complex split, complex merge or rename
   \item \code{before}: list of name of the admin1(s) before the event
   \item \code{after}: list of name of the admin1(s) after the event
   \item \code{d.before}: details of the previous admin2 units in the admin1
   associated concerned by the complex event
   \item \code{a.before}: details of the admin2 units in the admin1
   associated concerned after the complex event
}
}
\keyword{datasets}
