% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_province}
\alias{la_province}
\title{Lao Provinces dictionary}
\format{An object of class \code{character} of length 1353.}
\usage{
data(la_province)
}
\description{
A named character vector containing the translation in English of the
provinces names in Laos.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
