% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_province_year}
\alias{la_province_year}
\title{Lao Provinces By Year}
\format{An object of class \code{list} of length 3.
 \itemize{
   \item \code{"1997-2006"}: character vector of the 18 provinces in English
   \item \code{"2006-2013"}: character vector of the 17 provinces in English
   \item \code{"2013-2020"}: character vector of the 18 provinces in English
}}
\usage{
data(la_province_year)
}
\description{
A list of charactor vector containing the English names of the provinces
ordered by year. This list contains for each of change, the name of the "new"
provinces in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(la_province_year)

# To have the names of the province, between 1997 and 2006:
la_province_year$`1997`

}
\keyword{datasets}
