% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_year_admin1.R
\name{list_year_admin1}
\alias{list_year_admin1}
\title{Creates a lists of admin1 names by year range}
\usage{
list_year_admin1(country, hash, history_lst, from = "1960",
  to = "2020")
}
\arguments{
\item{country}{string character country name.}

\item{hash}{named vector used to translate Gadm names in English withot accent}

\item{history_lst}{list of event (see **_history object and
\code{make_history} function)}

\item{from}{year in character or numeric}

\item{to}{year in character or numeric}
}
\value{
a list of named vector, the names are the year of event.
}
\description{
From a vector of admin1 name downloaded from GADM, create a list, by year of
event (from \code{history_lst}) and by a time ramge (\code{from} and
\code{to} parameters), of admin1 names by year of change.
}
\examples{
library(dictionary)
la_admin1_year <- list_year_admin1("Laos", la_admin1, la_history,
                                   from = "1997")
}
