% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_history.R
\name{make_history}
\alias{make_history}
\title{Creates a list of event for a country}
\usage{
make_history(file, hash, d.hash)
}
\arguments{
\item{file}{txt file containing the information (try should be written in a
specific format, see \code{details})}

\item{hash}{named vector to translate the name of the admin1 geographic unit}

\item{d.hash}{named vector to translate the name of the admin2 geographic
unit}
}
\description{
From a text file (see prerequisite), make a list of a list of 4 or 6 elements
for each event (see Details).
}
\details{
From a text file (see prerequisite), make a list of event (list of 4 or 6
elements):
\enumerate{
   \item 'year': date of event in character,
   \item 'event': character either split, merge or rename,
   \item 'before': name of the admin1(s) before the event in a list and
   \item 'after': name of the admin1(s) after the event in a list
   \item 'd.before' : name of the admin2s concerned by the event
(only for complex event)
   \item 'd.after' : name of the admin2s concerned by the event
(only for complex event)
}
\cr\cr
In the file .txt, each event should be written in a specific format:
\code{"In DATE, ADMIN1(s) (ADMIN2(s)) EVENT in ADMIN1(s)."}. \cr
ADMIN1(s): if multiple, separated by ";" \cr
If the details of the split/merge event is available at admin2 level, the
admin2 should be written in (), separated by "," and the event should be
written complex EVENT. \cr
DATE: written by year as "YYYY" or can be a full date written as "YYYY-mm_dd"
\cr
EVENT: choose betwrepreen: split(s)/merge/rename(s) \cr
For example: In 1992, Hau Giang splits in Can Tho; Soc Trang. \cr
For example: In 2013, Vientiane (Longsan, Xaysomboun, Phun, Hom);
                      Xiengkhuang (Thathon) complex splits in Xaisomboun.
}
\examples{
\dontrun{
 library(dictionary)
 make_history("path/to/file.txt", vn_admin1)
}
}
