% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pattern.R
\name{match_pattern}
\alias{match_pattern}
\title{Compares one column against patterns}
\usage{
match_pattern(df, colname, lst_pattern, strict = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{colname}{a character string specifying the colonne to compare against
the list of pattern.}

\item{lst_pattern}{a named list of pattern, should contains one or multiple
vector. Each slot should be named.}

\item{strict}{a boolean to indicate if the character string should match only
some value in the vector (FALSE) or all values (TRUE) on the list of pattern.}
}
\value{
\code{match_pattern} In case of a matching, the function will return
a character vector, if not, it will return NULL
}
\description{
\code{match_pattern} compares one column of a data frame against a list of
pattern.
}
\details{
\code{lst_pattern} can contains multiple vector of pattern. The column
selected is compared to each vector contained in the list pattern and
return the name of the slot with the vector without any difference.
}
\examples{

# To look at the year of expression of the province in Vietnam from the
# imhen data sets (meteorological data)
library(imhen)

df <- imhen::meteo_intagg_2008_2017 # get data for the years 2008-2017
# Allows to check the spatial expression is corresponding to the year
match_pattern(df, "province", vn_admin1_year)

}
