% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces}
\alias{provinces}
\title{Provinces dictionary}
\format{An object of class \code{character} of length 429.}
\usage{
data(provinces)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the provinces names in Vietnamese.
}
\keyword{datasets}
