% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEA_country}
\alias{SEA_country}
\title{SEA Countries dictionary}
\format{An object of class \code{character} of length 42.}
\usage{
data(SEA_country)
}
\description{
A named character vector containing the translation in English of the
country names of South East Asia.
}
\keyword{datasets}
