% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{th_admin1}
\alias{th_admin1}
\title{Thai admin1 unit dictionary}
\format{An object of class \code{character} of length 3275.}
\usage{
data(th_admin1)
}
\description{
A named character vector containing the translation in English of the
admin1 unit names in Thailand.
}
\keyword{datasets}
