% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{th_admin1_year}
\alias{th_admin1_year}
\title{Thai admin1 unit By Year}
\format{An object of class \code{list} of length 7.
 \itemize{
   \item \code{"1967-1972"}: character vector of the 71 admin1 unit in English
   \item \code{"1972-1977"}: character vector of the 71 admin1 unit in English
   \item \code{"1977-1981"}: character vector of the 72 admin1 unit in English
   \item \code{"1981-1982"}: character vector of the 72 admin1 unit in English
   \item \code{"1982-1993"}: character vector of the 73 admin1 unit in English
   \item \code{"1993-2011"}: character vector of the 76 admin1 unit in English
   \item \code{"2011-2020"}: character vector of the 77 admin1 unit in English
}}
\usage{
data(th_admin1_year)
}
\description{
A list of charactor vector containing the English names of the admin1 unit
ordered by year. This list contains for each of change, the name of the "new"
admin1 unit in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(th_admin1_year)

# To have the names of the admin1, between 1993 and 2011:
th_admin1_year$`1993`

}
\keyword{datasets}
