% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{th_province}
\alias{th_province}
\title{Thai Provinces dictionary}
\format{An object of class \code{character} of length 3029.}
\usage{
data(th_province)
}
\description{
A named character vector containing the translation in English of the
provinces names in Thailand.
}
\keyword{datasets}
