% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{th_province_year}
\alias{th_province_year}
\title{Thai Provinces By Year}
\format{An object of class \code{list} of length 7.
 \itemize{
   \item \code{"1967-1972"}: character vector of the 71 provinces in English
   \item \code{"1972-1977"}: character vector of the 71 provinces in English
   \item \code{"1977-1981"}: character vector of the 72 provinces in English
   \item \code{"1981-1982"}: character vector of the 72 provinces in English
   \item \code{"1982-1993"}: character vector of the 73 provinces in English
   \item \code{"1993-2011"}: character vector of the 76 provinces in English
   \item \code{"2011-2020"}: character vector of the 77 provinces in English
}}
\usage{
data(th_province_year)
}
\description{
A list of charactor vector containing the English names of the provinces
ordered by year. This list contains for each of change, the name of the "new"
provinces in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(th_province_year)

# To have the names of the province, between 1993 and 2011:
th_province_year$`1993`

}
\keyword{datasets}
