% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Function to translate vector in English without accent}
\usage{
translate(vect, country, level)
}
\arguments{
\item{vect}{a string chracter vector.}

\item{country}{a character name of a country.}

\item{level}{a numeric, 1 for admin1 (province), 2 for admin2 (district),
3 for city/commune.}
}
\value{
string vector
}
\description{
Translates string vector in English without accent using dictionary created
and stored in the package by country and level of administrative boundaries.
}
\examples{
# to translate province name of Vietnam in English
translate(c("AnGiang", "Ha Noi"), "Vietnam", 1)
}
