% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_admin1}
\alias{vn_admin1}
\title{Vietnamese admin1 unit dictionary}
\format{An object of class \code{character} of length 5539.}
\usage{
data(vn_admin1)
}
\description{
A named character vector containing the translation in English of the
admin1 unit names in Vietnam.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
