% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_admin1_year}
\alias{vn_admin1_year}
\title{Vietnamese admin1 unit By Year}
\format{An object of class \code{list} of length 7.
 \itemize{
   \item \code{"1979-1990"}: character vector of the 40 admin1 unit in English
   \item \code{"1990-1991"}: character vector of the 44 admin1 unit in English
   \item \code{"1991-1992"}: character vector of the 45 admin1 unit in English
   \item \code{"1992-1997"}: character vector of the 53 admin1 unit in English
   \item \code{"1997-2004"}: character vector of the 61 admin1 unit in English
   \item \code{"2004-2008"}: character vector of the 64 admin1 unit in English
   \item \code{"2008-2020"}: character vector of the 63 admin1 unit in English
}}
\usage{
data(vn_admin1_year)
}
\description{
A list of charactor vector containing the English names of the admin1 unit
ordered by year.
}
\details{
One specificity of Vietnam is the splits of administrative admin1 unit during
its history, starting from 40 admin1 unit in 1980 and ending with 63 admin1 unit
since 2008. This list contains for each of change, the name of the "new"
admin1 unit in English.
}
\examples{
# To have the year of each events of splitting/merging:
names(vn_admin1_year)

# To have the names of the admin1, between 1992 and 1997:
vn_admin1_year$`1992`

}
\keyword{datasets}
