% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_admin2}
\alias{vn_admin2}
\title{Vietnamese admin2 units dictionary}
\format{An object of class \code{character} of length 21158.}
\usage{
data(vn_admin2)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the admin2 units names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
