% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_district}
\alias{vn_district}
\title{Vietnamese Districts dictionary}
\format{An object of class \code{character} of length 21158.}
\usage{
data(vn_district)
}
\description{
A names charactor vector containing the translation from Vietnamese to
English of the districts names in Vietnamese.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
