% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_history}
\alias{vn_history}
\title{Vietnamese Provinces Boundaries History (since 1979)}
\format{An object of class \code{list} of length 24.}
\usage{
data(vn_history)
}
\description{
List of 24 lists containing 4 elements:
 \itemize{
   \item \code{year}: date of event in character in format "YYYY-mm-dd"
   \item \code{event}: character vector of one object either split, merge or
   rename
   \item \code{before}: list of name of the province(s) before the event
   \item \code{after}: list of name of the province(s) after the event
}
}
\keyword{datasets}
