% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vn_province}
\alias{vn_province}
\title{Vietnamese Provinces dictionary}
\format{An object of class \code{character} of length 5013.}
\usage{
data(vn_province)
}
\description{
A named character vector containing the translation in English of the
provinces names in Vietnam.
The dictionary is encoded in UNICODE
}
\keyword{datasets}
