% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Retrieve the dictionary definition of an English word}
\usage{
define(word)
}
\arguments{
\item{word}{The word you wish to define}
}
\value{
A data.frame with as many rows as distinct meanings, and columns for
    word, phonetic, phonetics, origin, and meanings
}
\description{
Retrieve the dictionary definition of an English word
}
\examples{
\dontrun{
  # Get the definition of "hello"
  define("hello") # Returns a 1 row data.frame

  # Homonyms (words with >1 meaning)
  define("bark") # Returns a 4 row data.frame

  # Unknown words will error
  define("sdfsdfsdfsdfsdf")

}

}
