% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary-package.R, R/dictionary.R
\docType{package}
\name{dictionary}
\alias{dictionary}
\alias{dictionary-package}
\title{dictionary: Dictionary of Vietnamese admin1 Names}
\usage{
dictionary(countryname, level, add_dict = NULL, force = FALSE)
}
\arguments{
\item{countryname}{string character country name}

\item{level}{numeric level of administrative boundaries (1: admin1,
2: admin2), should be either 1 or 2}

\item{add_dict}{which contains at least one column \code{translate} (english
without accent) and one or multiple columns containing the tag `var` in the
column names (all this columns will be used to create the possible variable)}

\item{force}{boolean to force the gadm downlaoding if trouble}
}
\description{
This package contains a dictionary for Vietnamese admin1 names, admin2
names and admin3 names, useful to ensure consistency between different data
 packages using vietnamese admin1, districts and/or admin3.

From Gadm and Geonames creates a UNICODE dictionary, names vector, with the
translation in English without accent.
}
\details{
You can create dictionary for different level: 1 (admin1) and 2 (admin2),
from the data of GADM and Geonames but you can also add costum value by using
 the argument \code{add_dict}. It is also possible to use only the parameter
 \code{add_dict} to generate custom dictionary.\cr
The variables created are express in upper, lower cases, with or without
space or "_" or with capital etters at the beginning of each word and
expressed it in UNICODE. \cr
Source : gadm: https://gadm.org \cr
Source: geonames: http://download.geonames.org/export/dump/
}
\examples{
dict_vn <- dictionary("Vietnam", 1)
head(dict_vn)
}
