% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{package}
\name{dictionary}
\alias{dictionary}
\alias{dictionary-package}
\title{dictionary: Dictionary of Vietnamese Provinces Names}
\description{
This package contains a dictionary for Vietnamese provinces names, districts
names and communes names, useful to ensure consistency between different data
 packages using vietnamese provinces, districts and/or communes.
}
