% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pattern.R
\name{match_pattern}
\alias{match_pattern}
\title{Compares one column against patterns}
\usage{
match_pattern(df, colname, lst_pattern)
}
\arguments{
\item{df}{a data frame.}

\item{colname}{a character string specifying the colonne to compare against
the list of pattern.}

\item{lst_pattern}{a named list of pattern, should contains one or multiple
vector. Each slot should be named.}
}
\value{
\code{match_pattern} returns a character vector.
}
\description{
\code{match_pattern} compares one column of a data frame against a list of
pattern.
}
\details{
\code{lst_pattern} can contains multiple vector of pattern. The column
selected is compared to each vector contained in the list pattern and
return the name of the slot with the vector without any difference.
}
\examples{

# To look at the year of expression of the province in Vietnam from a
# gdpm data frame (epidemiologic data from Vietnam)
library(gdpm)

df <- getid(dengue) # get data for dengue for Vietnam
match_pattern(df, "province", vn_province_year)


}
