% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translates Vector in English Without Accent}
\usage{
translate(vect, hash, country = NULL, level = NULL)
}
\arguments{
\item{vect}{a string chracter vector.}

\item{hash}{a named UNICODE vector containing the translation}

\item{country}{a character name of a country.}

\item{level}{a numeric, 1 for admin1 (province), 2 for admin2 (district),
3 for city/commune.}
}
\value{
String vector of the same length as input in \code{vect} argument
}
\description{
Translates string vector in English without accent using dictionary created
and stored in the package by country and level of administrative boundaries.
}
\details{
The functions accepts two types of entry for translation. \cr
1. You can use the \code{country} and \code{level} information to translate
vector, but only for Vietnam, Cambodia and Thailand.
2 . You can directly input a named vector in UNICODE containing the
translation.
If the argument \code{hash} is NULL, the vector will be return encoded in
UNICODE.
}
\examples{
# to translate province name of Vietnam in English
translate(c("AnGiang", "Ha Noi"), vn_province)
# or
translate(c("AnGiang", "Ha Noi"), country = "Vietnam", level = 1)
}
